Add test for Shared library defined custom MaterialX surface shader nodes
The MateiralXFilter essentially replaces the terminal/surfaceshader
sdr node with a new one holding the materialX generated glslfx file.
Nodes using primvars like texture coordinates indicate they are using
these primvars through their sdr node metadata. So if we had a
custom surface shader node that was using texture coordinates,
we would loose the metadata with the primvar information, when
we created the new sdr node. The fix for this was to initialize the
new sdr node with the original sdr metadata so we don't loose
this information. This fix was added to the MaterialXFilter in CL 2370558.

This change adds a test to catch this case and re-names and
re-organizes the CustomNodes test since we will be adding a
LocalCustomNodes test shortly.

(Internal change: 2370793)