Add WorkTaskGraph to migrate parallel exec engine code that relies on the deprecated low-level tbb::task interface.
WorkTaskGraph implements a base task class that supports spawning
additional work from within the execution of a task, continuation passing,
recycling of task resources, and scheduler bypass. This change also includes
an implementation of tbb::zero_allocator, which was removed in the oneTBB
migration as an oversight.

This enables pxr to build with oneTBB and continues to support task usage
with older TBB versions.

(Internal change: 2370806)
(Internal change: 2370808)