Add version support to usda files.
Until now all usda layers have been version "1.0", even when new
features have been added that cannot be parsed by older readers. This
change adds a mechanism for code that is writing to the file to request
that an upgraded version be used because of some feature that is being
written.

In addition, the version of a layer when it is read is preserved when
the layer is saved. Specifically, if SdfLayer::SaveToFile is used then
the input version will be propagated to the output. If WriteToFile is
used then the file will start at the default output version. In either
case, the version may be upgraded if more advanced features are
discovered when writing the layer.

The pieces of this change are:
  * The CrateFile::Version class has been extracted to SdfFileVersion
    and enhanced.

  * The Sdf_TextOutput class now has RequestWriteVersionUpgrade that can
    be called by writing code to upgrade the version of the output file.

  * Sdf_TextOutput now also tracks and updates the version number in the
    layer header.

  * SdfUsdaFileFormat has been extended with minimum and maximum input
    and output file versions. The maximum input or output file version
    is currently 1.1 for testing purposes even though we never write
    real version 1.1 files. The first file format change will become
    version 1.1 and we'll upgrade from there.

  * Layers read from usda files now use SdfUsdaData as the layer's data
    member. SdfUsdaData retains the version of the input file so that
    it can be used when the file is saved. Files that are exported or
    written to other destinations start with the default version and get
    upgraded by RequestWriteVersionUpgrade as needed while writing.

(Internal change: 2371085)