This change removes the compatability shim placed in stage.cpp in 2342828. The result is that finer grained change notifications are boradcast to consumers of UsdNotice::ObjectsChanged notices.
In order to ensure the correct notices are boradcast the following modifications are required:

PcpChanges::DidChangeSpecs considers layer stacks which are being unmuted or inserted when searching for the node of a changed spec.  This prevents some added inert prims from ending up in PcpCacheChanges::didChangePrims and causing USD to treat them as resyncs.

USDStage::_AddAffectedStagePaths, like above needs to consider layer stacks which are being unmuted or inserted when adding paths to ChangedPath maps.  Failure to do this can result in some paths not being property sorted into the correct maps and UsdStage specific change processing will not take place.

When processing LayersDidChange notifications we need to ensure all changelists generated as a result of sublayer operations are passed to UsdStage::_ProcessChangeLists

(Internal change: 2371115)