usdviewq: Initialize ArResolver outside of "open stage" timer
This restores previous behavior that was inadvertently changed
by change 2370231. In theory the time to initialize the resolver
is a fixed cost that is generally negligible when loading
non-trivial assets, but we do have performance tests that
benchmark very small stages and if we just left this as-is
those numbers would get noisier.

(Internal change: 2371173)