Make Zipfile.cpp validate content length for zip32
Adds validations to zipFIle write out so that we can make sure we're only writing valid Zip32 files.
This prevents several forms of data corruption, and allows for early exits.

Testing of this functionality is facilitated by the addition of an EnvSetting which controls the maximum size of the archive.  The description of this setting notes it is intented for testing purposes only.

This is based on PR #3618 by @dgovil.

Closes #3618

(Internal change: 2371220)