[hdSt] Add support for camera exposure compensation.
Camera exposure is represented as a scalar multiplier on incoming radiance ("linear exposure scale"), calculated from an exponential "exposure" adjustment as well as potentially fStop/iso/etc in the case of UsdGeomCamera.  This change adds Storm support for exposure compensation, by multiplying the result of "surfaceShader" calls (calculating lit and shaded 3d pixels) by the exposure factor.  Note that exposure compensation is not applied to some other sources of color like "repr" color or selection highlight colors, which are considered non-physical colors.

Since exposure is only applied to the rendering process, exposure compensation of framebuffer contents that will be composited with hydra needs to be done as a separate process.

(Internal change: 2371431)