Fix build_usd.py installing USD libs into '--instDir' on macOS
Before this change, an invalid string comparison on macOS was made when mount paths like:
  '/private'
  '/System/Volumes/Data'
were returned as path prefixes in os.getcwd(), causing USD to be installed into the same directory as the '--inst' (dependencies) folder, instead of the install_dir folder.

With this change, the assignment of the directory to install each component into can be made explicitly when calling RunCMake() with its new installDir kwarg.

USD is now correctly installed into install_dir on macOS when '--instDir' is used on the command line to specify an alternate directory for third party dependencies.

(Internal change: 2371536)