Handle addition/removal of mesh light status in the mesh light resolver.
With new workflows for applying the mesh light API to targeted geometry,
prims can now start or stop being mesh lights on the fly. E.g. they can
gain or lose a true value for "isLight" in their data source. Previously
this didn't seem to be accounted for in the mesh light resolver's
_PrimsDirtied().

(Internal change: 2371553)