[hdPrman] Reorder geom subset primvar inheritence
Previously we copied the parent mesh's primvars first, then tried to convert the subset's primvars on top of them. This meant primvar conversion's routines intended for avoiding duplicate primvars on the same prim would kick in and ignore anything that was already on the mesh when setting the subset's primvars. That's not really how HdPrman_ConvertPrimvars() is intended to be used.

Instead, we start by converting the subset's primvars, then we use RtPrimVarList::Inherit() to let the subset's primvars inherit non-conflicting primvars from the mesh. This is the correct pattern, because it helps avoid coding errors like this one, although it costs extra memory. It's also the only way to achieve what we want in the case of geom subsets.

Also note that Inherit() and Update() have their own quirks with RtPrimVarList -- if the destination primvar list's detail sizes don't match the source's, primvars with the respective detail settings will be skipped.

I would have added an assertion at the top of HdPrman_ConvertPrimvars() that the given RtPrimVarList/RtParamList is empty, but we are still passing non-empty param lists all over the place right now to save on memory and due to how we separately process the points primvar. Just be aware that pattern is less transparent than using the explicit Inherit() and Update() methods.

(Internal change: 2371597)