exec: Handle input recompilation of multiple input keys.
Computations can register multiple inputs of the same name and type. In
compilation, each input key is resolved individually, and the outputs
resolved from all matching input keys are connected to a single VdfInput of
that name and type.

Input recompilation was not handling this case correctly. It would only
recompile the first input key with the same name as the input, instead of
recompiling all input keys with the same name. This change corrects
Exec_NodeRecompilationInfo and Exec_InputRecompilationTask to properly
recompile an input if it sources values from multiple input keys.

(Internal change: 2371653)