Teach TsSpline to use tangent algorithms
This change adds tangent algorithms to knots. There are separate tangent
algorithms for the pre- and post-tangents. Whenever the knots in a
spline changes, it recomputes the tangents on the knots based on their
tangent algorithms.

The known tangent algorithms are:
  * None: do not change the user supplied tangents (see exception below)
  * Custom: look in the knot's custom data for "preTanAlgorithm" and/or
    "postTanAlgorithm" keys. If the value is a string that names an
    algorithm that is either built in or supplied via some TBD
    registration method, then the specified algorithm will be invoked to
    compute the tangent.
  * AutoEase: the "Auto Ease" algorithm from Maya/AnimX.

The binary writing code for crate files has been extended to write
splines that contain tangent algorithms as version 0.13.0.

The text writing code for usda files has been extended to write splines
that contain tangent algorithms as version 1.1.

(Internal change: 2371698)