Fix meshes disappearing when unapplying the MeshLightAPI.
This can be relevant when changing the collection membership expression
on a standalone mesh light model or a PLP with a mesh light source. If a
mesh loses mesh light status, we need to be sure to send an appropriate
notification to indicate that the mesh should be considered a mesh again.

(Internal change: 2372022)