build: Fix issue with missing TBB::tbb target in pxrConfig.cmake
Change 2369739 updated USD to specify library dependencies on TBB
with the more modern TBB::tbb target. This caused the targets in
USD's package config file to specify TBB::tbb as an interface link
library instead of the absolute path to the .so/.dll.

This caused problems when trying to link other cmake-based projects
to the USD package for USD builds using legacy TBB. In this case,
the TBB::tbb target (typically) is created by USD internally and
isn't available to clients, which caused "target was not found"
errors.

This fix recreates the TBB::tbb target in pxrConfig.cmake to ensure
its available in the above case. This should yield the same behavior
as before the change mentioned above.

(Internal change: 2372031)