OpenExec: Implement plugin attribute computations, including dispatched attribute computations.
Plugin attribute computations are analogous to plugin prim computations, but
whereas prim computations are found by requesting a computation with a provider
prim, attribute computations are found on provider attributes. In particular,
exec registrations for a given schema can register a plugin computation that
will be found on attributes of a given name, on prims that have the schema.

Dispatched attribute computations are analogous to dispatched prim computations,
except again the provider in this case is an attribute. Assuming all of the same
criteria that apply to dispatched prim computations are met (including matching
schemas and that the requesting input specifically requests dispatched
computations), a dispatched attribute computation will be found on any
attribute. There is no need or ability to specify the name of the attributes to
which a dispatched attribute computation applies.

(Internal change: 2372342)
(Internal change: 2372354)