hdPrmanXpuLoader: Fix link issues on Linux and macOS
The hdPrmanXpuLoader plugin currently links against hdPrmanLoader,
which is also a plugin. However, the build system assumes that
plugins are leaf libraries with no dependents and does not set
up an RPATH entry pointing to the plugin install directory.
This leads to an error when trying to switch to RenderMan XPU
in usdview: loading hdPrmanXpuLoader fails because it can't
find its hdPrmanLoader dependency at runtime.

This fix works around the issue by forcing the addition of the
plugin install directory to hdPrmanXpuLoader's RPATH. This is
intended to be a short term patch while we explore other ways
to structure these loader plugins.

(Internal change: 2372879)