OpenExec: Fix a problem in callback registration where it was easy to register a callback where the result value would be ignored.
E.g., no value would be produced by the following computation:

    self.PrimComputation(_tokens->primComputation)
        .Callback<std::string>(+[](const VdfContext &) { return "a string"; });

Note that the callback is a function pointer whose return type is convertible to
std::string, but is not actually a std::string. Previously, the function pointer
would be implicitly cast to ExecCallbackFn, which is std::function<void (const
class VdfContext &)>, because std::function is happy to cast away the return
type.

In order to minimize risk of users defining callbacks with thread safety hazards
or side effects that could thwart exec invalidation and caching, we no longer
accept capturing lambdas as callbacks. We now require all callbacks to be passed
as function pointers.

(Internal change: 2372887)