[hdPrman] Conform PxrEnvDayLight behavior to documentation
PxrEnvDayLight rendered through hdPrman results in inverted east-west axis when using date/time/place to set sun position, and setting "month" to 0 and using "sunDirection" instead additionally moves the zenith to +Z instead of +Y.

This is ultimately due to a mismatch between RenderMan documentation for PxrEnvDayLight and its implementation. The implementation in RenderMan needs to remain the way it is for now, so it is up to the bridge products (like hdPrman) to compensate. We have not been making the correct compensation in hdPrman.

This change adds the required special compensation needed to conform PxrEnvDayLight behavior to the documentation and make it consistent with the other RenderMan DCC bridges. This includes introducing an east-west flip of the dome's orientation, and a rotation-and-flip of "sunOrientation" from the documented coordinate space (+Y to zenith, [implied] -Z to north, [implied] right-handed) into RenderMan's expected coordinate space (+Z to zenith, +X to north, left-handed).

Fixes #3675

(Internal change: 2372910)