Support for disabling complex macro TF_DECLARE_PUBLIC_TOKENS for intellisense
The boost free version of the preprocessor macros on MSVC can cause issues with Intellisense
processing, particularly around `TF_DECLARE_PUBLIC_TOKENS`.  This change suppresses that macro
when processed by intellisense to prevent it from spending major processing power attempting
to interpret the macro.

https://github.com/microsoft/vscode-cpptools/issues/13205#issuecomment-2732102126

Adds a check for the traditional preprocessor so it should be fine if / when the conformant preprocessor is enabled:

https://github.com/PixarAnimationStudios/OpenUSD/pull/3565

Closes #3732

(Internal change: 2373382)