Fix swizzle node handling
Change 2371696 resulted in all nodes to have their nodedef name
set. However we do not want to do that for these swizzle node
replacements since it results in a nodecategory/nodedefname
mismatch that MaterialX will complain about.

Before the above mentioned change, setting the nodeDefString was
guarded by a mxNode->getNodeDef() check. This meant that these
swizzle node replacements did not have a nodedef string set and
MaterialX was happy. This change adds that check back before
setting the nodeDefString.

This also adds a defensive check in the materialXFilter for when we try
and get the sdrNode for a materialX deprecated node in
_ConnectPrimvarNodesToTerminalNode()

Fixes #3738

(Internal change: 2373506)