Avoid coding error with benign duplicate connectable behavior registration
If we are trying to register a behavior for a primTypeId which already has a
behavior registered, we should not allow it, but if multiple threads tried to
register the same behavior (like a null behavior for caching purpose, its
benign and we silently allow it)

More background:
When a call to _GetBehaviorForPrimTypeId results in a behavior not being
found in the cache or in any of the apiSchemas associated with
primTypeId, we associate a null behavior with this primTypeId, so as to
do a fast lookup in any further GetBehavior query.

With this change, we check if the behavior being registered again in
another thread is same, then we silenly allow it, but if a different
behavior is being registered for a primTypeId, we still throw a coding
error.

Fixes #3547

(Internal change: 2373804)