[hd] Fix SampleExtComputationInput() in backend emulation for non-varying computation inputs.
This fixes failures in `HdExtComputationUtils::SampleComputedPrimvarValues()` we encountered when evaluating skinning computations in Hydra 2

If the computation input was not time-sampled, the return value would be 0 causing callers to ignore the time sample of 0.0. This could then cause `HdExtComputationUtils::SampleComputedPrimvarValues()` to fail while executing the computation due to the non-varying inputs having no values.

This was the same issue that 33dbdb15 fixed for `SamplePrimvar()`

Closes #3704

(Internal change: 2373889)