Autodesk: [hdSt] Make the order of MaterialX texture uniforms consistent across platforms
The generated Storm MaterialX shaders differs between platforms due to an std::unordered_map.
After the Storm material network was anonymized, this attribute no longer needs to be a map
at all. This change updates the attribute type to a Vector.

Note that testHdStMaterialXShaderGen cannot be enabled on more platforms until USD is upgraded
to MaterialX v1.38.4 because of a similar problem fixed on the MaterialX side: https://github.com/AcademySoftwareFoundation/MaterialX/pull/2462

Closes #3728

(Internal change: 2373954)