[usdSkelImaging] Block authored normals to match UsdImaging 1.0 behavior
This means that the normals need to be computed post-skinning (by the render delegate), rather than using the undeformed normals from the rest geometry, and matches the UsdImaging 1.0 behavior and addresses bug #3699.

Note that the open PRs (e.g. #3664) to add support for normals in the skinning computation would supersede this.

Closes #3700

(Internal change: 2374067)