[hd] Add custom translator hook point from dirtybits to locator set (and vice versa) for Rprims.
This PR enables registration of custom translators between dirtybits and locator sets for Rprims. It mimics the logic used for Sprims, except that if the prim type is not found there will be no invalidation.

This was discussed in more detail here: https://forum.aousd.org/t/scene-index-invalidation-of-custom-prim-types/2544 and fixes issue #3694.

Closes #3722

(Internal change: 2374165)