sdf,usd: Add limits metadata field for attributes
UsdAttribute now has a dictionary-valued `limits` field which can house an
arbitrary number of sub-dictionaries, each of which can contain a minimum
and/or maximum value, along with custom keys. "Hard" and "soft" subdicts
are provided for, which encode strict and loose limits, respectively. But
clients may define their own subdictionaries as well.

Access API is provided by the UsdAttributeLimits class, which reads and
writes to a single subdict in the larger limits dict.

(Internal change: 2374314)