pcp: Fix simplified specializes regression with variants
In certain cases where a specializes arc was introduced inside
a variant, the process of adding implied specializes arcs up to
the root of the prim index would be cut short causing valid
opinions to be omitted. This worked correctly prior to the
introduction of simplied specializes.

For example, consider the case where the variant node at
/A/B{x=y} introduces a specializes arc to /A/C. That arc would
be immediately propagated to the root of the graph and have
other composition tasks enqueued -- particularly the implied
class task. Pcp would try to add the corresponding implied arc
to the parent of the variant node at /A/B. The computed implied
specializes path would be /A/C, so an inert placeholder node would
be added at the path and would then be immediately propagated to
the root. However, since there was already a propagated specializes
arc at /A/C, no node would be added and no tasks enqueued, which
would cause the implied class process to stop.

This change detects this situation and ensures an implied class
task is always enqueued to ensure the process continues. While
this fixes the reported regression, it introduces a separate
regression in an internal test. This second regression involves
a much more complicated composition structure that we think is
less likely, so I'm landing this fix first to address the more
common case for the 25.08 release. This change will be reverted
afterwards while work on a more complete fix is underway.

See #3750

(Internal change: 2374321)