[hd] Fix race condition in HdDataSourceLegacyPrim.
UsdImagingDelegate::GetPrimvarDescriptors() is not threadsafe
because it may invoke UpdateForTime().  A comment says it does
this under the assumption that it is being invoked during render
index population, where the external code will only invoke this on
independent prims in parallel.

This is unfortunately not the case when a scene index is invoking
methods on HdDataSourceLegacyPrim from multiple threads.

HdDataSourceLegacyPrim now uses a mutex to serialize its
population of the primvars data source.

(Internal change: 2374664)