exec: Add Exec_TaskCycleDetector
It is reasonable that exec may try to compile a computation that depends on
itself. For example, computation A declares an input on computation B, while
computation B declares an input on computation A.

In such cases, the Exec_CompilationTasks would directly or indirectly form
a dependency cycle, causing one or more tasks to never complete because a
task is stuck waiting on itself. As a result, exec compilation would hang.

This change adds an Exec_TaskCycleDetector that tracks the state of all
Exec_CompilationTasks and detects when compilation is about to get stuck in
a task cycle. When a task cycle is detected, the process is aborted to
prevent the hang.

(Internal change: 2374847)