[hdx] Add baseline hints Hgi_GL and Hgi_Vulkan, depending on whether env var HGI_ENABLE_VULKAN is enabled. We use this env var to give testHdxPickTarget a separate set of baseline images for Vulkan.
The two edge selection highlighting test cases in testHdxPickTarget previously
would fail when run with HgiVulkan, seemingly due to differences in the shader
dFdx and dFdy functions. The Vulkan result, though hard to see in the test due
to highlight color, is actually much less jaggy than the GL result. Because
this result is distinct, confined to a specific feature (edge selection
highlighting), and improved with Vulkan, keeping it as an alternate baseline
seemed appropriate.

(Internal change: 2374922)