OpenExec: Implement Metadata input node invalidation
Currently, all field changes that aren't otherwise specifically handled by exec
change processing are passed along as possible metadata changes. If we haven't
compiled a metadata input node for a given field, the change will be ignored, so
processing non-metadata fields in this way is correct. There may be some
performance cost, however. We *could* filter up-front by querying if a given
field is a metadata field, but it's not clear if that would be faster. It's
something we can come back and investigate later.

(Internal change: 2374977)