Implement TsSpline::Breakdown and CanBreakdown
Breakdown inserts a new knot into the spline at the provided time such
that the spline changes as little as possible. This involves splitting
one of the segments of the spline into two segments.

It's trivially easy to split value-block, held, or linear segments. It
is almost as easy to split Hermite curves, we just have to create a new
knot with the value and derivative of that timem on the spline. Bezier
curves are slightly more involved.

Bezier curves are parameterized curves with each segment evaluated over
a [0..1] range. In order to split the curve at a given time, we first
have to solve for the u parameter that yields that time. The solving
code that provides that is in eval.cpp so the breakdown code has been
added there.

The function _FindBezierParameter has been split out of _EvalBezier and
does that solve. _FilterZeroes has been simplified, made less
susceptible to round-off errors, and is now spelled with the correct
number of 'e's (_FilterZeros).

When there is looping, one or both of the knots we're interpolating may
be echoes of another knot. The _ReplaceXXX() routines have been
extended to return optional flags indicating whether the updated knot
data is generated or part of the spline's data.

Breakdown determines where in the spline it is being asked to operate
(pre- or post-extrapolation or between knots) and generates a new knot
accordingly.

CanBreakdown uses the same code as Breakdown but exits early as soon as
it has decided whether it can insert a knot or not.

Ts_SplineData has had a GetDataAsDouble method that returns a
Ts_TypedKnotData<double> regardless of the spline's data type. Now it
also has a SetKnotFromDouble method that does the inverse and stores a
Ts_TypedKnotData<double> even for other data types.

New spline methods have been wrapped to python. CanBreakdown uses
TfPyAnnotatedBoolResult to return both the bool and the reason why it
may be false.

(Internal change: 2375179)