Fix expansion of varying and indexed-vertex primvars for pinned curves.
HdsiPinnedCurveExpandingSceneIndex is used to translate BasisCurve
primitives using the Hydra wrap mode "pinned" into BasisCurves using the
wrap mode "nonperiodic", in order to be usable by render delegates that
don't natively support "pinned" (specifically: HdPrman).

Two related issues were discovered: first, the input-size validation
that the scene index did on varying primvar inputs was incorrect,
causing it to skip expanding them.  Once that was corrected, it was
discovered that the expansion being applied was wrong.  Both situations
resulted in the primvar data presented to HdPrman being the wrong size
(first too small, then too large), which HdPrman would then omit.

These issues both stem from a misunderstanding of how many input values
are appropriate for "pinned" curves; and what impact that number has on
how the data gets expanded.  This change corrects that, which also
simplifies the expansion process as there was a branch of code to
support combinations of values that are not valid arrangements with the
corrected interpretation of how many primvar values there should be for
different curve lengths.

While fixing this a latent bug was also discovered where indexed vertex
primvars on curves that use a topology index buffer would have their
indices expanded.  This is wrong when the index buffer in the topology
is already being expanded; the result is doubly-expanded primvar values.
This change fixes that oversight and updates the test as well for the correct behavior.

(Internal change: 2375860)