OpenExec: Add static assert to ensure that equality comparison is defined for all registered exec value types.
Equaility comparison was already required for types that are registered with the
ExecTypeRegistry. But before this change, the compiler errors were long and hard
to parse (at least for gcc), since the compiler ends up emitting errors for many
equality comparison operators that _don't_ work for the registered type. With
this change, the compiler errors are still long, but the static assert ensures
that near the top, there's an easy-to-read explanation of what went wrong.

(Internal change: 2376147)