[hydraPrimvarPruning] Second try at moving primvar pruning scene index to run extremely late, just prior to Riley.
The idea with primvar pruning is that, for RenderMan specifically, we
can avoid sending it primvar data that we know is unused within the
rendering process.  We do this based on specific knowledge of the
RenderMan concepts we use, such as material networks, light filters
that read associated primvars, and so on.

The primvar pruning scene index must not, however, interfere with
anything upstream of the renderer that happens to make use of
primvars for proper functioning.  The primvar pruning scene index
should not interfere with this process, and also should not be
on the hook to be informed about every possible consumer of
primvars -- just those within RenderMan.

A few fixes required for this:

- HdPrman's render delegate now pulls on SdrRegistry at init time.
It seems generally better to bring-up important subsystems
like this at specific known times.

(Internal change: 2376273)