[hdx] Fix skydome rendering.
The skydome task is used as a clear in some of our pipelines (which makes sense, since it's placed at the far plane), but even though it was writing to all of the pixels it wasn't running an actual clear, which can trip up some of our framebuffer setups (e.g. anything using stencil).

(Internal change: 2376412)