[hdEmbree] Initial UsdLux reference implementation
This modifies the hdEmbree example plugin to do direct lighting so as to provide a reference implementation of the expected behaviour for UsdLux.

If no lights are present in the stage, or if direct lighting is turned off (default) via the render setting "enableLighting", the old ambient occlusion path will be used.

The intent for this implementation is to be spec-compliant and readable, more than scalable.

Closes #3186

(Internal change: 2376413)
(Internal change: 2376439)