vdf: VdfLinearMap is now VdfConnectorMap
VdfLinearMap was primarily used to store inputs and outputs belonging
to a node.  There are a small number of other uses but the long-term
goal is to specialize this data structure for the input/output case.
To this end, rename VdfLinearMap to VdfConnectorMap and restrict its
keys to TfToken and it's mapped type to pointers.

Furthermore, because the users of VdfLinearMap managed the lifetimes
of the pointed-to objects in a way that was coupled to their presence
in the map, VdfConnectorMap takes ownership of its mapped values.

(Internal change: 2376688)