OpenExec: Implement constant input parameters
A constant input parameter provides a constant value as an input to a
computation. This kind of input isn't necessarily useful when used with a
self-contained computation definition. But it becomes useful for more
complicated registrations, where one piece of code registers a callback that
configures its evaluation-time behavior based on an input value and a separate
piece of code registers a constant input that selects the desired behavior.

This can happen:
- When computation definitions are assembled programatically by parameterized
  registration code that is called to register various versions of a
  computation, possibly for multiple schemas
- When computation definitions are composed from registrations made for
  different schemas on the same prim (support for composed computation
  definitions is still TBD in OpenExec)
- When computation registration is configured (also TBD), allowing registrations
  for a single schema to be dynamic, depending on metadata values that drive the
  configuration process

(Internal change: 2377437)