Make dome light use cubemaps.
This change switches Storm from using latlong textures for dome lights to
instead use cubemap textures.  A cubemap is generated from the latlong dome
light texture, and cubemaps are used for the irradiance and prefilter textures.
The BRDF texture remains a 2D texture since it is a LUT.

This addresses some visual artifacts that could arise when using latlong
textures, particularly when sampling along the poles (though this also depends
on the contents of the dome light texture and roughness level of the surface).
Otherwise the overall appearance is very similar to using latlong textures,
though it will vary slightly.  In particular, it can cause some shifting of
highlights, and preserve more variance at low roughness levels because we have
six small cube face textures instead of one small texture.

Fixes #2776

(Internal change: 2377735)