Update sdf text file parsing for empty splines and extra white space.
Change the text file parser to allow a spline's value to consist of 0 or
more spline items rather than 1 or more. This allows completely empty
splines to be read.

Change the text file and crateFile writers for splines to write a
spline even if it does not have any knots.

Update the testSdfParsing test file 220_splines.usda to verify that
empty splines can be read and written and that extra white space near
preValue & value pairs is parsed correctly.

Update the testUsdSplines and testUsdSplinesCpp tests to correctly
verify the round tripping of empty splines.

Fixes #3772
Fixes #3773

(Internal change: 2377812)