pcp: Fix/workaround for specializes issue with duplicate arcs
Simplified specializes introduced a regression in an asset
setup that caused Pcp to imply two specializes arcs with
different origins but pointing to the same site to the same
parent node in the prim index.

Prior to simplified specializes, the arc with the stronger
origin happened to be added to the prim index first. Indexing
would then skip adding the second arc because it would duplicate
the arc that had already been added. Simplified specializes
wound up flipping this so that the arc with the weaker origin
would be added first, and the arc with the stronger origin
would be skipped. This changed the strength ordering of
specializes opinions, since origin strength is part of the
overall strength ordering computation.

This change addresses this issue by changing the order in
which composition evalutes implied class tasks. Previously,
tasks for nodes that were added _later_ to the prim index
would be handled before nodes that were added earlier to
ensure that tasks for child nodes were processed before
their parents. Now, tasks for nodes that are not related
ancestrally are handled in strength order to ensure that
implied arcs with stronger origins are processed first.

I think there is a more general solution that involves changing
how composition deals with duplicate arcs, but that's a much
larger change. For now, this is a targeted fix for this
specific issue.

(Internal change: 2378155)