vt/sdf: Remove the ability for VtArrayEdit to represent dense VtArrays.  This was done so that the composition operator always operated on the single VtArrayEdit type, like:
VtArrayEdit * VtArrayEdit -> VtArrayEdit

Where '*' is the composition operator and VtArray implicitly converts to a
dense VtArrayEdit.

But this turns out to be more complicated than just letting the composition
operator also work on VtArray.  For example, code that deals with VtArrays held
in VtValues is burdened with treating VtValues holding VtArrayEdits
representing dense arrays the same as VtValues holding ordinary VtArrays.

By removing this ability, now a VtArrayEdit always only represents incremental
edits, and only VtArray represents dense arrays.  So we have:

VtArrayEdit * VtArrayEdit -> VtArrayEdit
VtArrayEdit * VtArray     -> VtArray

The VtArray * VtArrayEdit -> VtArray case is well-defined but it's trivially
the left hand side so we leave it out for now.  It can be added if needed.

We continue to write the 'isDense' field as always false to crate files to
avoid bumping the version number, under the assumption that since this feature
is not complete yet, real assets with dense arrays in VtArrayEdit instances
have not been created.

(Internal change: 2378293)