HdMergingSceneIndex: ensure prims exists if its path is a scene root or an ancestor of a scene root - even if the input scene doesn't have a prim at the scene root.
Note that the HdLegacyPrimSceneIndex is empty (not even having a prim at the absolute root path) but is added to the HdMergingSceneIndex (with sceneRoot being the absolute root path).

When removing the UsdImaging scene index from the HdMergingSceneIndex, it still has the empty HdLegacyPrimSceneIndex at the absolute root path. Thus, the correct invalidation is to resync the absolute root path and remove all its children that came from the UsdImaging scene index.

This invalidation was happening before change list 2377732 because it checked whether a prim at the pseudo root exists by checking both whether GetPrim returns a non-null data source and whether GetChildPrimPaths is non-empty. Change list 2377732 removed the latter.

This change fixes it.

(Internal change: 2378386)